Move ELTJobManifest definition to typing module
